requirements.txt
django>=4.0.0,<=4.1.0
djangorestframework
pyyml
requests
django-cors-headers
psycopg2
django-environ
uritemplate